refactor like button component
