+++ src/components/index.js
@@ -3,5 +3,6 @@ export { default as FooterBar } from './footer-bar/footer-bar';
 export { default as HeroElement } from './hero-element/hero-element';
 export { default as RatingBar } from './rating-bar/rating-bar';
 export { default as RestoList } from './resto-list/resto-list';
+export { default as LikeButton } from './like-button/like-button';
 export { default as LoadingIndicator } from './loading-indicator/loading-indicator';
 export { default as NotificationToast } from './notification-toast/notification-toast';
