+++ src/components/like-button/like-button.scss
@@ -1,0 +1,20 @@
+@import 'src/styles/config';
+@import 'src/styles/helpers';
+
+like-button {
+    display: block;
+    
+    button {
+        cursor: pointer;
+        border-radius: 50%;
+        font-size: 1.5rem;
+        height: 44px;
+        width: 44px;
+        border: none;
+        background: $color-regular-button;
+    }
+    
+    i.favorite {
+        color: $white;
+    }
+}
